requirements.txt
ultralyticsplus
transformers
pillow